requirements/base.txt
# SHA1:a9dde048f1ee1f00586264d726d0e89f16e56183
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-e file:.
    # via -r requirements/base.in
apsw==3.38.5.post1
    # via shillelagh
attrs==23.1.0
    # via
    #   cattrs
    #   requests-cache
cattrs==23.1.2
    # via requests-cache
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
greenlet==2.0.2
    # via
    #   shillelagh
    #   sqlalchemy
idna==3.3
    # via requests
packaging==23.0
    # via shillelagh
platformdirs==3.11.0
    # via requests-cache
python-dateutil==2.8.2
    # via shillelagh
requests==2.31.0
    # via
    #   requests-cache
    #   shillelagh
requests-cache==1.1.0
    # via shillelagh
six==1.16.0
    # via
    #   python-dateutil
    #   url-normalize
sqlalchemy==1.4.39
    # via shillelagh
typing-extensions==4.3.0
    # via shillelagh
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.10
    # via
    #   requests
    #   requests-cache
